refactor audit subviews
